Refactored and added a bookmark component!
