+++ src/components/Content/List.jsx
@@ -9,17 +9,7 @@ const List = ({ cards }) => {
     <>
       {
         cards.map((card) => (
-          <Card
-            key={card.id}
-            id={card.id}
-            bookmark={card.bookmark}
-            poster_path={card.poster_path}
-            name={card.name}
-            title={card.title}
-            release_date={card.release_date}
-            first_air_date={card.first_air_date}
-            adult={card.adult}
-            media_type={card.media_type}
+          <Card key={card.id} {...card}
           />
         ))
       }
